+++ src/components/index.ts
@@ -1,6 +1,6 @@
 import { Editor, useEditor } from "@tiptap/vue-3";
 import { editorExtensions } from "@tiptapify/components/editorExtensions";
-import { toolbarSections } from "@tiptapify/types/toolbarSections";
+import { toolbarSections } from "@tiptapify/types/toolbarTypes";
 import { ShallowRef } from "vue";
 
 export function getTiptapEditor (
@@ -11,11 +11,10 @@ export function getTiptapEditor (
   onUpdate: Function = () => {}
 ): ShallowRef<Editor | undefined> {
   const extensions = editorExtensions(placeholder, slashCommands, customExtensions)
-  const editor: ShallowRef<Editor | undefined> = useEditor({
+
+  return useEditor({
     content,
     extensions,
     onUpdate: ({ editor }) => onUpdate()
   })
-
-  return editor
 }