+++ src/extensions/components/formatExtra/quote/Button.vue
@@ -5,6 +5,12 @@ import { Editor } from "@tiptap/vue-3";
 import BtnIcon from "@tiptapify/components/UI/BtnIcon.vue";
 import { inject, Ref } from "vue";
 
+import defaults from '@tiptapify/constants/defaults'
+
+defineProps({
+  variantBtn: { type: String, default: defaults.variantBtn }
+})
+
 const editor = inject('tiptapifyEditor') as Ref<Editor>
 
 const { t } = inject('tiptapifyI18n') as any
@@ -15,6 +21,7 @@ const { t } = inject('tiptapifyI18n') as any
   <VBtn
       :color="editor.isActive('blockquote') ? 'primary' : ''"
       :disabled="!editor.can().chain().focus().toggleBlockquote().run()"
+      :variant="variantBtn"
       @click="editor.commands.toggleBlockquote()"
       size="32"
   >