+++ src/extensions/components/media/image/Button.vue
@@ -6,6 +6,12 @@ import BtnIcon from "@tiptapify/components/UI/BtnIcon.vue";
 import ImageDialog from "@tiptapify/extensions/components/media/image/ImageDialog.vue";
 import { inject, Ref } from "vue";
 
+import defaults from '@tiptapify/constants/defaults'
+
+defineProps({
+  variantBtn: { type: String, default: defaults.variantBtn }
+})
+
 const editor = inject('tiptapifyEditor') as Ref<Editor>
 
 const { t } = inject('tiptapifyI18n') as any
@@ -15,6 +21,7 @@ const { t } = inject('tiptapifyI18n') as any
 <template>
   <VBtn
       :color="editor.isActive('image') ? 'primary' : ''"
+      :variant="variantBtn"
       @click="editor.commands.showTiptapifyImage()"
       size="32"
   >